service-workers/service-worker/navigation-preload/empty-preload-response-body.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Navigation Preload empty response body</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.sub.js"></script>
<script>
promise_test(async t => {
  const SCRIPT = 'resources/empty-preload-response-body-worker.js';
  const SCOPE = 'resources/empty-preload-response-body-scope.html';
  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  const worker = registration.installing;
  await wait_for_state(t, worker, 'activated');
  const frame = await with_iframe(SCOPE);

  add_completion_callback(async() => {
    if (registration)
      await registration.unregister();
    if (frame)
      frame.remove();
  })

  assert_equals(frame.contentDocument.body.textContent, '[]');
}, 'Navigation Preload empty response body.');

</script>
